Refactored to adapt to the refactor of the merge_mps code
