+++ src/cpp/benders/merge_master_mps/CMakeLists.txt
@@ -9,7 +9,6 @@
 add_library(merge_master_mps_core)
 target_sources(merge_master_mps_core PRIVATE
         ${CMAKE_CURRENT_SOURCE_DIR}/include/antares-xpansion/benders/merge_master_mps/MergeMasterMPS.h
-        ${CMAKE_CURRENT_SOURCE_DIR}/include/antares-xpansion/benders/merge_master_mps/StandardLp.h
         ${CMAKE_CURRENT_SOURCE_DIR}/include/antares-xpansion/benders/merge_master_mps/MasterCoupling.h
 
         ${CMAKE_CURRENT_SOURCE_DIR}/MergeMasterMPS.cpp
@@ -25,6 +24,7 @@ target_include_directories(merge_master_mps_core
 target_link_libraries(merge_master_mps_core
         PUBLIC
         antaresXpansion::benders_core
+        antaresXpansion::merge_mps_core
         antaresXpansion::helpers
         antaresXpansion::logger_lib
         antaresXpansion::factories
