+++ src/cpp/benders/merge_master_mps/MergeMasterMPS.cpp
@@ -1,12 +1,13 @@
 #include "antares-xpansion/benders/merge_master_mps/MergeMasterMPS.h"
-#include "antares-xpansion/benders/merge_master_mps/StandardLp.h"
 #include "antares-xpansion/benders/merge_master_mps/MasterCoupling.h"
+
 
 #include <filesystem>
 #include <utility>
 #include <numeric>
 
 #include "antares-xpansion/benders/benders_core/CouplingMapGenerator.h"
+#include "antares-xpansion/benders/merge_mps/StandardLp.h"
 #include "antares-xpansion/helpers/Timer.h"
 
 MergeMasterMPS::MergeMasterMPS(MergeMasterMPSOptions options,
